ImageModelTrainer/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
imageio==2.34.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.5.22
Werkzeug==3.0.3
m2cgen